requirements.txt
Pillow==6.0.0
PySocks==1.7.0
PyYAML==5.1.1
